new front page with new + upload buttons
