+++ src/router/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Route, Switch } from 'react-router-dom';
-import FormBuilder from '../views/FormBuilder';
+import FrontPage from '../views/FrontPage';
 import { TreeContextProvider } from '../store/treeStore/treeStore';
 import Login from '../views/Login';
 import Code from '../views/Code';
@@ -10,7 +10,7 @@ export default function Routes(): JSX.Element {
         <Switch>
             <Route path="/" exact>
                 <TreeContextProvider>
-                    <FormBuilder />
+                    <FrontPage />
                 </TreeContextProvider>
             </Route>
             <Route path="/login" exact>
